pyproject.toml
[project]
name = "yb-csv-importer"
version = "0.1.0"
description = "A standard YugabyteDB CSV import and benchmarking tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-yugabytedb-binary>=2.9.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
